pyproject.toml
[project]
name = "wav2sleep"
version = "0.1.0"
description = "End-to-end sleep staging from raw physiological signals"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface_hub>=0.20.0",
    "hydra-core>=1.3.2",
    "lightning==2.1.0", # Pin because the MLFlow logger is broken.
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "pyarrow>=19.0.0",
    "pyedflib>=0.1.42",
    "omegaconf>=2.3.0",
    "torch>=2.7.1",
    "mlflow<3",
    "seaborn>=0.13.2",
    "numba>=0.60.0",
    "hydra-useful-callbacks",
    "submitit==1.5.3",
    "hydra-submitit-launcher",
    "setuptools<81",
]


[tool.ruff]
line-length = 120
fix = true
exclude = ["docs"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = ["S", "I", "F"]
# Skip unused variable rules: https://www.flake8rules.com/rules/F841.html
# Loose imports: https://www.flake8rules.com/rules/F403.html
ignore = ["F841", "S311", "S108", "F403", "S506", "S101", "S603"]

[tool.ruff.lint.isort]
case-sensitive = true

[tool.uv.sources]
hydra-useful-callbacks = { git = "https://github.com/joncarter1/hydra_useful_callbacks.git" }
hydra-submitit-launcher = { git = "https://github.com/facebookresearch/hydra.git", subdirectory = "plugins/hydra_submitit_launcher", tag = "v1.3.2" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hydra-colorlog>=1.2.0",
    "ipykernel>=6.29.5",
    "ipympl>=0.9.7",
    "pytest>=8.4.1",
    "ray[default]>=2.48.0",
    "ruff>=0.12.2",
]
